perf(edgeless): re-render only related parts
